BME280 Pressure, Temperature and Humidity Sensor
The i2c addressable BME280, about the size of a grain of rice, comes ready-mounted on PCBs from SparkFun and Adafruit.  The Adafruit Python driver was written for the Arduino and cannot be directly used on the Microbit.  The Adafruit driver is , in turn, based on the Bosch sample code and echoes its anomalous technique of converting 12-bit signed values into Python integers.  The sensor is most sensitive, draws very little current and is worth exploring - the pressure sensor permits altitude determination with meaningful precision and accuracy (100Pa resolution).  The code is a work in progress; it tests the Microbit's memory limit and there are issues I would very much like to discuss with anyone who is interested.